requirements.txt
transformers>=4.36.0
accelerate>=0.24.0
bitsandbytes>=0.41.1
sentence-transformers>=2.2.2
chromadb>=0.4.14
faiss-cpu>=1.7.4
pinecone-client>=2.2.2
scikit-learn>=1.3.0
optuna>=3.3.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
replicate>=0.20.0
openai>=1.3.0